+++ streaming/java/streaming-common/src/main/java/io/ray/streaming/common/serializer/proxy/SerializerProxy.java
@@ -1,0 +1,13 @@
+package io.ray.streaming.common.serializer.proxy;
+
+import io.ray.streaming.common.serializer.KryoUtils;
+import org.nustaq.serialization.FSTConfiguration;
+
+/** use for loading different serializer */
+public abstract class SerializerProxy {
+
+  public abstract byte[] encode(Object obj);
+
+  public abstract Object decode(byte[] obj);
+}
+
