+++ streaming/java/streaming-state/src/main/java/io/ray/streaming/state/SnapshotResult.java
@@ -8,7 +8,9 @@ public class SnapshotResult {
 
   public boolean result;
 
-  public SnapshotResult() {};
+  public SnapshotResult() {
+
+  }
 
   public SnapshotResult(long snapshotId, boolean result) {
     this.snapshotId = snapshotId;
